Fix warnings that prevent compiling
With the current "SRV_IDL_DRSBind", there are two comparisons made between signed and unsigned LONGs (lines 2361 and 2369).
This is because FIELD_OFFSET returns a LONG, but the value you are comparing it with (cb) is a ULONG. In visual studio the project settings for mimikatz will treat warnings as errors and won't compile.

To fix it, I've changed size from a ULONG to a LONG (line 2355) because it will only every be assigned a LONG value, and I cast the initial FIELD_OFFSET comparison as a ULONG (line 2362). This fixes both warnings and allows the project to compile with no warnings.