Remove check for '.' in domain name for dcsync
The code in dcsync was checking to make sure that the domain name provided either by the user or by the result of an API call was an FQDN. To do this, it was checking that a period ('.') was present in the name. Often FQDNs contain periods, however I was on a gig recently that had at least one domain that did not have a period in the FQDN, and as a result the calls to dcsync were failing.

This PR updates mimikatz so that it no longer checks for the period in the FQDN, and assumes that the user knows what they're doing!